+++ src/redux/epics.js
@@ -20,6 +20,7 @@ import * as timezoneEpics from './modules/entities/timezones/epics';
 import * as businessHoursEpics from './modules/entities/businessHours/epics';
 import * as fetchReportingEventsEpics from './modules/entities/reportingEvents/epics';
 import * as tenantsEpics from './modules/entities/tenants/epics';
+import * as reasonsEpics from './modules/entities/reasons/epics';
 
 export const rootEpic = combineEpics(
   ...Object.values(brandingEpics),
@@ -42,5 +43,6 @@ export const rootEpic = combineEpics(
   ...Object.values(timezoneEpics),
   ...Object.values(businessHoursEpics),
   ...Object.values(fetchReportingEventsEpics),
-  ...Object.values(tenantsEpics)
+  ...Object.values(tenantsEpics),
+  ...Object.values(reasonsEpics)
 );